Use docs/README.md + script to generate extension-specific README.md
This relies on the fact that GitHub will look for READMEs either in .github, or
in the root or in docs, in this order. So moving the template-specific README to docs
makes for it to be overtaken by anything that sits in the root.

python3 script/set_extension_name.py my_own_extension_name my_own_function_name
 * copies docs/NEXT_README.md into README.md
 * changes <extension_name> occurrences in my_own_extension_name
 * changes quack occurrences in my_own_function_name